+++ front-end/src/app/reports/f3x/f3x-routing.module.ts
@@ -4,9 +4,9 @@ import { CreateF3XStep1Component } from './create-workflow/create-f3x-step1.comp
 import { ReportSummaryComponent } from './report-summary/report-summary.component';
 import { ReportDetailedSummaryComponent } from './report-detailed-summary/report-detailed-summary.component';
 import { ReportResolver } from 'app/shared/resolvers/report.resolver';
-import { ReportLevelMemoComponent } from '../report-level-memo/report-level-memo.component';
+import { ReportLevelMemoComponent } from '../shared/report-level-memo/report-level-memo.component';
 import { TestDotFecComponent } from './test-dot-fec-workflow/test-dot-fec.component';
-import { PrintPreviewComponent } from '../print-preview/print-preview.component';
+import { PrintPreviewComponent } from 'app/reports/shared/print-preview/print-preview.component';
 import { CashOnHandComponent } from './create-workflow/cash-on-hand.component';
 import { CashOnHandGuard } from 'app/shared/guards/cash-on-hand.guard';
 import { ReportIsEditableGuard } from '../../shared/guards/report-is-editable.guard';
@@ -81,7 +81,10 @@ const routes: Routes = [
     component: ReportLevelMemoComponent,
     canActivate: [ReportIsEditableGuard],
     resolve: { report: ReportResolver, sidebar: SidebarStateResolver },
-    data: { sidebarSection: ReportSidebarSection.REVIEW },
+    data: {
+      sidebarSection: ReportSidebarSection.REVIEW,
+      getNextUrl: (report?: Report) => '/reports/f3x/submit/step1/' + report?.id,
+    },
     runGuardsAndResolvers: 'always',
   },
   {
