+++ front-end/src/app/reports/reports.module.ts
@@ -21,14 +21,14 @@ import { TooltipModule } from 'primeng/tooltip';
 
 import { DialogModule } from 'primeng/dialog';
 import { InputNumberModule } from 'primeng/inputnumber';
-import { SharedModule } from '../../app/shared/shared.module';
+import { SharedModule } from 'app/shared/shared.module';
 import { AppSelectButtonComponent } from '../shared/components/app-selectbutton.component';
 import { CashOnHandComponent } from './f3x/create-workflow/cash-on-hand.component';
 import { CreateF3XStep1Component } from './f3x/create-workflow/create-f3x-step1.component';
 import { ReportDetailedSummaryComponent } from './f3x/report-detailed-summary/report-detailed-summary.component';
-import { ReportLevelMemoComponent } from './report-level-memo/report-level-memo.component';
+import { ReportLevelMemoComponent } from './shared/report-level-memo/report-level-memo.component';
 import { ReportSummaryComponent } from './f3x/report-summary/report-summary.component';
-import { PrintPreviewComponent } from './print-preview/print-preview.component';
+import { PrintPreviewComponent } from './shared/print-preview/print-preview.component';
 import { TestDotFecComponent } from './f3x/test-dot-fec-workflow/test-dot-fec.component';
 import { FormTypeDialogComponent } from './form-type-dialog/form-type-dialog.component';
 import { ReportListComponent } from './report-list/report-list.component';
